# tacocloud/tacocloud/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: TacoCloud API
  version: 1.0.0
  description: Public v1 contract. Equivalent /api routes are legacy aliases and return Deprecation and Link headers.
servers:
  - url: http://localhost:8080/api/v1
security:
  - basicAuth: []
...
components:
  securitySchemes:
    basicAuth: {type: http, scheme: basic}
  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema: {type: string, maxLength: 64, pattern: '^[A-Za-z0-9._:-]+$'}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: {type: string, minLength: 8, maxLength: 128, pattern: '^[A-Za-z0-9._:-]+$'}
...
  responses:
    BadRequest: {description: Invalid request, content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}}
    Forbidden: {description: Access denied, content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}}
    NotFound: {description: Resource not found, content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}}
    Conflict: {description: Conflict, content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}}
  schemas:
    ApiViolation:
      type: object
      properties: {field: {type: string}, message: {type: string}}
    ApiProblem:
      type: object
      required: [type, title, status, detail, code, violations]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string}
        code: {type: string, example: IDEMPOTENCY_KEY_REUSED}
        violations: {type: array, items: {$ref: '#/components/schemas/ApiViolation'}}
...
    TacoRequest:
      type: object
      required: [name, ingredientIds]
      properties:
        name: {type: string, maxLength: 50, example: Veggie Supreme}
        ingredientIds: {type: array, minItems: 2, uniqueItems: true, items: {type: string}}
...
    OrderLineRequest:
      type: object
      required: [taco, quantity]
      properties:
        taco: {$ref: '#/components/schemas/TacoRequest'}
        quantity: {type: integer, minimum: 1, maximum: 10}
    OrderCreateRequest:
      type: object
      required: [deliveryName, deliveryStreet, deliveryCity, deliveryState, deliveryZip, paymentMethodId, items]
      properties:
        deliveryName: {type: string, maxLength: 80}
        deliveryStreet: {type: string, maxLength: 120}
        deliveryCity: {type: string, maxLength: 80}
        deliveryState: {type: string, maxLength: 50}
        deliveryZip: {type: string, minLength: 3, maxLength: 12}
        paymentMethodId: {type: string, maxLength: 64, description: Tokenized payment method identifier}
        couponCode: {type: string, maxLength: 40}
        items: {type: array, minItems: 1, maxItems: 20, items: {$ref: '#/components/schemas/OrderLineRequest'}}
    OrderResponse:
      type: object
      description: Safe public representation without authentication, raw payment or internal User data.
      properties:
        id: {type: string}
        placedAt: {type: string, format: date-time}
        deliveryName: {type: string}
        deliveryStreet: {type: string}
        deliveryCity: {type: string}
        deliveryState: {type: string}
        deliveryZip: {type: string}
        status: {type: string, enum: [CREATED, ACCEPTED, PREPARING, READY, OUT_FOR_DELIVERY, DELIVERED, CANCELLED]}
        subtotalBeforeDiscount: {type: number}
        discountAmount: {type: number}
        total: {type: number}
        currency: {type: string, example: USD}
        items: {type: array, items: {type: object}}
    AnnouncementRequest:
      type: object
      required: [text, severity, expiresAt]
      properties:
        text: {type: string, minLength: 1, maxLength: 500}
        severity: {type: string, enum: [INFO, WARNING, CRITICAL]}
        expiresAt: {type: string, format: date-time}
...
paths:
...
  /orders:
    get: {tags: [Orders], summary: List own orders or all for ADMIN, responses: {'200': {description: Orders}}}
    post:
      tags: [Orders]
      summary: Create an idempotent order
      parameters: [{$ref: '#/components/parameters/CorrelationId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/OrderCreateRequest'}}}}
      responses:
        '201': {description: Created or replayed, headers: {X-Correlation-Id: {schema: {type: string}}}, content: {application/json: {schema: {$ref: '#/components/schemas/OrderResponse'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '409': {$ref: '#/components/responses/Conflict'}
  /orders/{orderId}:
    parameters: [{name: orderId, in: path, required: true, schema: {type: string}}]
    put: {tags: [Orders], summary: Replace owned order, responses: {'200': {description: Updated}, '404': {$ref: '#/components/responses/NotFound'}}}
    patch: {tags: [Orders], summary: Patch delivery fields, responses: {'200': {description: Updated}, '404': {$ref: '#/components/responses/NotFound'}}}
    delete: {tags: [Orders], summary: Delete owned order, responses: {'204': {description: Deleted}, '404': {$ref: '#/components/responses/NotFound'}}}
...
  /users/me/orders/{orderId}:
    get: {tags: [History], summary: Get own historical order, parameters: [{name: orderId, in: path, required: true, schema: {type: string}}], responses: {'200': {description: Order}, '404': {$ref: '#/components/responses/NotFound'}}}
...
  /admin/announcements:
    get: {tags: [Operations], summary: List active announcements (ADMIN), responses: {'200': {description: Active announcements}, '403': {$ref: '#/components/responses/Forbidden'}}}
    post: {tags: [Operations], summary: Create announcement (ADMIN), requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AnnouncementRequest'}}}}, responses: {'201': {description: Created}, '400': {$ref: '#/components/responses/BadRequest'}, '409': {$ref: '#/components/responses/Conflict'}}}
  /admin/announcements/{id}:
    delete: {tags: [Operations], summary: Delete announcement (ADMIN), parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: {'204': {description: Deleted}, '404': {$ref: '#/components/responses/NotFound'}}}